+++ src/index.ts
@@ -8,6 +8,9 @@ export { default as State } from "./accounts/State";
 export { default as Margin } from "./accounts/margin/Margin";
 export { default as Cache } from "./accounts/Cache";
 export { default as Control } from "./accounts/Control";
+export { default as MarginsCluster } from "./accounts/margin/MarginsCluster";
+
+export { decodeEvent } from "./utils/events";
 
 export { ZoMarket, Orderbook, ZoOpenOrders } from "./zoDex/zoMarket";
 export { EVENT_QUEUE_HEADER, decodeEventsSince } from "./zoDex/queue";