+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.0]
+- Added Power Perps Support
+- Removed cache as param for MarginWeb3 Functions
+- Added liquidation calculations to MarginWeb3
+- Added margin exists function to check if margin exists
+- Added state & cache subscriptions
+- 
 
 - Program: Add FillOrKill order type. Acts as an IoC which fails if not completely filled ([#67](https://github.com/01protocol/zo-client/pull/67))